Move `FileOptionsProvider` to `:cli`
